+++ activerecord/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Deprecate `.reorder(nil)` with `.first` / `.first!` taking non-deterministic result.
+
+    To continue taking non-deterministic result, use `.take` / `.take!` instead.
+
+    *Ryuta Kamizono*
+
 *   Ensure custom PK types are casted in through reflection queries.
 
     *Gannon McGibbon*
@@ -77,5 +83,6 @@
 *   Allow generated `create_table` migrations to include or skip timestamps.
 
     *Michael Duchemin*
+
 
 Please check [6-0-stable](https://github.com/rails/rails/blob/6-0-stable/activerecord/CHANGELOG.md) for previous changes.